[INTERNAL] ResourceHandler: added on-the-fly compilation of CSS/JSON
Similar like for the Grunt based infrastructure the Java server now also
supports the on-the-fly compilation of the library(-RTL).css and
library-parameters.json. The responses are cached and the invloved
resources are watched so in case of a modification of an involved
resources the cached for that compilation results is invalidated and
recompiled. 

With that change the client-side LessSupport is obsolete for the OpenUI5
testsuite.

Change-Id: Ic62fb489dc6b7587872eb22255ee6400c92aba7f