[FIX] core.Popup: Fixed Messed Up Event Queue in IE
- If using a scroll bar of a ListBox within an autoclose Popup IE messes
up the events. Plus it fires several blur events when clicking on the
scroll-bar. At some point all these timeouts within the Popup are
totally out of sync and a close call occurs and closes the Popup and its
children
- Now in the close-function a check is made if the close-function was
called by the autoclose delayedCall and if the activeElement is still
within the Popup -> then the closing will be cancelled.
- Plus if the delayedCall-id is still set within the closing function it
has to be cleaned to prevent any further confusion

BCS: 0120025231 0000942225 2014
Change-Id: I030df32253184ab4110fbbe17a2e0e6338361c49