[FIX] sap.m.IconTabBar: Unwanted spacing
Removing 1px unwanted spacing just below the selected tab arrow. The
issue exists in Chrome and IE, under Firefox this 1px spacing does not
exist, because FF sets larger line-height (with 1px) than IE and Chrome.
To unify the bahavior and remove the spacing I add line-height with 1px
larger than Chrome and IE sets and change the line-height of the text
only usage to cope with the line-height change; The changes effect only
the bottom border position, the text stay in the same place;

Change-Id: I7b0a12474219482b026e07399868c7ac6caf3d47
CSS: 0120031469 0000739799 2014