[INTERNAL] sap.ui.core coding should run on IE8 whenever possible.
Mid term, we should consider to add polyfills for the ECMA5 stuff
that is missing in IE8 and migrate the code to the ECMA5 methods.

As this is a significant shift in the coding guidelines, it requires 
a bit more thinking (and discussion and testing). As a short term
solution, jQuery substitute methods need to be used.

'Most wanted' candidates for polyfills:
- Array.isArray
- Array.prototype.every
- Array.prototype.filter
- Array.prototype.forEach
- Array.prototype.indexOf
- Array.prototype.map
- Array.prototype.some
- Date.now
- Function.prototype.bind
- Object.keys
- String.prototype.trim()

Note: some packaged thirdparty components are known not to support 
IE8: blanket, d3, handlebars, less.

Change-Id: I9c13000d77e3d9f669fc4b07124f954c5c93fe51