[FIX] ODataModel - correct error handling for metadata requests
When a new model tries to access shared metadata from the metadata
cache, and finds metadata which is broken because of a previously failed
request, it will now automatically trigger a refresh. When this happens,
and the request fails again, the subsequently created model will hence
also receive a metadataFailed event and can react accordingly.

Change-Id: Ide27d27d85c55ba27637fdf9dd8037fd9d87510f