Fixes http://github.com/SAP/openui5/issues/207
Added the ariaDescribedBy and ariaLabelledBy associations to the
sap.m.Button control. Also updated the explored page for buttons to use
the new associations. I added a new disabled button to the explored sample
to ensure that bringing over the code from the desktop version of Button,
which used a slightly different handling of the disabled property, still
worked correctly.