[FEATURE] Component: usage of manifest for custom metadata
Instead of polluting the metadata of the Component with additional
custom entries we now support the definition of a Component manifest
which is aligned with the W3C standard for web application manifests.
The namespace for SAPUI5 entries is below "sap.ui5" where most of the
entries of the former Component metadata will be moved into. The legacy
Component metadata will be still supported for compatibility reasons but
internally the Component converts the legacy metadata into the new
manifest structure so the with the new public API getManifest the
necessary data can be accessed in a common way. 

The old specific metadata API is now deprecated and only the new generic
functions getManifest and getManifestEntry are public. This has been
changed because of the JSON structure has been standardized and is
reliable. In addition it would be hard to provide a public API for all
special entries here. For compatibility reasons the old public API is
still available and returns the values as before which is done by
transforming the manifest values into the legacy format. But all legacy
functions are now deprecated and must not be used anymore. 

In addition all occurences of jQuery.each have been replaces with the
respective for loop.

Increased test coverage for external, inline, missing or empty
metadata/manifest files to make sure the all options are working even
when modifying or improving the implementation later. This should also
ensure the fail safety of the conversion concept.

Minor comment regarding models creation: for the old legacy metadata the
models creation was implemented but as it was experimental and mentioned
to be changed it will not work for now for the new manifest metadata.
This feature will be enabled for the manifest later. But it still works
for the legacy metadata.

Change-Id: I040cc333834ec0aad4ec01c3ab374f39e073ab02