[FIX] Templating: clone the render fn and consider binding context
When using a TemplateControl as template for an aggregation in the
databinding framework it needs to clone the render function otherwise it
simply cannot render anymore.

The second issue fixed by that change is that the templating now also
considers the binding context of the root element when used inside an
aggregation. This allows to use the TemplateControl also with relative
bindings and not absolute ones.

Change-Id: Ie178510304f0f24c0da5ee6beb29a5a3aff727d6
(cherry picked from commit 84544c375d4921b8b113ffa42c1c73bc23e31ab2)
CR-Id: 012003146900008994792014