[FIX] Make LessSupport plugin work with new module 'value' of Core
The AMD definition of sap.ui.core.Core has been changed so that it 
only exposes the public API of the Core. The LessSupport plugin wasn't
adapted to this change and tests didn't reveal it. 

Also ensure that only the interface of the Core is exposed.

(cherry picked from commit ff2fd9aef6beff744fed91644cf3aabc6be58b5d)
Change-Id: Id38a1d1fe0dbcdcf8aefa7ebb3ce25a7f108a1c5