[FIX] sap.m.Input suggestionItemSelected event handling improvement
The suggestionItemSelected event is now fired before the update of the
DOM item and the setting of the input value.

This allows to process this event by application hooks and suggestion
providers before the input field is updated by the control.

If the value of the input field is changed in the registered event
handler, the value of the selected item is ignored and the custom value
is used instead.

BCP: 1482017037
Change-Id: If6bd186ead4343fa3a561d7f36e709c2a82d88df