[FEATURE] Accessibility: add lang attribute to HTML element for ARIA
In order that ARIA knows the language of the document it requires the
lang attribute on an HTML element. In case of UI5 applications the whole
document is normally in the same language so we can write the locale
information to the HTML element centrally.

Additionally the Core listens to the localization change event and
adopts the lang attribute properly.

Change-Id: I0dcf6c83f4c197989d4a40e25b05dd7c3e12bd36