[FIX] DateFormat: Better support 01010001 dates coming from ABAP
While it sounds like a date not relevant in real world applications, in
ABAP it is usual to create date ranges from 01010001 to 31129999. In
this case there was an issue when the date was formatted and parsed
again where the year was changed to 2001 instead. This has now been
fixed by formatting years less than 100 to 00XX, even if four digits are
note required in the date pattern.

BCP: 1482027172
Change-Id: I8c9fda63358ef231b993272e3a2b5f6085400d9c
CR-Id: 012006153200015086802014