[FIX] P13ColumnsPanel : fix for reset feature inside columns panel
- in case the smart table works without variant management the table
personalization dialog offers a reset button to reset all changes to the
original data. If this reset is executed the P13nColumnsPanel destroys
all of its items and columnsItems and create them new. During this
creation step all unselected table items (inside the columns panel of
the personalization dialog) will be made to invisible, which is wrong.
They shall be unselected instead.

Change-Id: I1a7ad305fd75c233cd4f08b430aa1c58b624cc83
CR-Id: 012006153200015135972014