[INTERNAL] sap.ui.core.Popup: Event target check
At some places in the code we are checking manually
if the target is the same as the popup dom node and
there are some other places where this check is not
even present. In the case of the MessageToast this
is a problem since there is only one dom node and
when checking only with jQuery.contains() it returns
false.

Change-Id: Ia5b69474f2d73f319885b34d3ec83a8b87adde6a
Signed-off-by: Deyan Yanakiev <deyan.yanakiev@sap.com>