[FIX] jQuery.sap.global: includeStyleSheet error callback for IE
Fixed the includeStyleSheet to trigger an error callback in case a
stylesheet cannot be loaded. Unfortunately we can provide this feature
only with a workaround where we count the amount of rules inside the
stylesheet and we assume that at least one rule is contained. This check
is only done for IE because other browsers behave correctly. This
specific behavior is also mentioned in the JSDoc for the error handler.

Change-Id: I002567f94d2390ae7287243cf466eb1f772df1ef
CSS: 0120061532 0001293129 2014
BCP: 1472011229