[FIX] core.Popup: Moved Check if Running in a Popup
- Moved the check if a Popup is running in another Popup before the
Popup is being rendered via 'this._$(true)'. Because when the Datepicker
and its Calendar are rendered the corresponding element will be focused
immediatelly. So the parent popup with the Datepicker's input field will
loose the focus and the focus will be set to the specified date. Since
the Calendar's popup wasn't added as child to the parent popup yet the
check in 'onFocusEvent' in 'Popup.js' isn't able to check the focus
properly.

CR-Id: 012006153200015137482014
CSS: 0120031469 0000743725 2014
Change-Id: I7389871aeba3d9b5e6cd1b386123254b34f9cac3