[FIX] ODataTreeBinding: Regression when binding to collection
There was a regression in the ODataTreeBinding, when it was bound to a
collection instead of a single entity. In 1.24 in this case all nodes
contained in the collection were displayed as root nodes, while in 1.26
the first node was assumed to be the root node and was automatically
hidden.

- Fix regression issue
- Avoid unnecessary requests for child nodes, if data was already
expanded
- Avoid multiple calls change event, before data is available

BCP: 1570000032
Change-Id: Ida49db3762a7c8e0afcda107c93190400e38c5b2