[FIX] sap.m.Carousel: Do not display unnecessary hover buttons
Currently rigthmost and leftmost hover buttons in a non looping carousel
are displayed with opacity 0. This leads to the issue that underlying
controls may react not correctly. The issue has been solved be setting
the css display property of unnecessary hover buttons to 'none'
  
Also, the page indicator is rendered only if the page displays more than
one page. Otherwise, the page indicator would be useless. 

Change-Id: I47ced42be7ae56e193a4272c8b94b97ecb4975d8
CSS: 1472020642