[INTERNAL] ManagedObject: allow declaration of special settings
Special settings are accepted by a class constructor and/or the
applySetting function for a class, but they are not properties,
aggregations, associations or events. 

The declaration of such settings helps generic functionality to 
distinguish really unknown settings from settings that are just 
not described in normal metadata. 

Special settings are still an internal feature.

Change-Id: I3405e3ea49235f82f3139f311ce2c6f9a7b61807