[FIX] LessSupport: differ between no last-modified header and not found
In case of e.g. CSS files are not available and the server doesn't
support last modified HTTP headers the less support misinterprets the
state and tries to load the CSS instead of using the LESS files. With
that fix the existing LESS files will be used when no CSS file is
available even if there is no last-modified header available.

Change-Id: Ia493ec03700f735065d2481049058eba8ea6fc16