[FIX] SelectDialog: remove from static UIArea before destroying dialog
The SelectDialog adds itself to the static UIArea on open. Because of
that, the automatic removal of the content from the UIArea as
implemented by the sap.ui.core.Popup no longer applies. The popup
removes the content only when it added it itself. 

As a result, a destroy() of the dialog invalidated the static UIArea
which in turn destroyed other instances of SelectDialog (as encountered
with test case SelectDialog.qunit.html).

Change-Id: Id3ed3cbd9aff379dcfdc602834966a7f2603732f