[INTERNAL][FIX] Table SelectDialog:  remove from static UIArea before destroying dialog
The SelectDialog adds itself to the static UIArea on open. Because of that, the automatic removal of the content from the UIArea as implemented by the sap.ui.core.Popup no longer applies. The popup removes the content only when it added it itself.
As a result, a destroy() of the dialog invalidated the static UIArea which in turn destroyed other instances of SelectDialog (as encountered with test case SelectDialog.qunit.html)
Related to change id: https://git.wdf.sap.corp:8080/#/c/765202

Change-Id: I11e43d44df490b290b1959a07d7e92f3ca0b5ebb