[FIX] sap.m.ViewSettingsDialog: remove from static UIArea before destroying dialog
The ViewSettingsDialog adds itself to the static UIArea on open. Because
of that, the automatic removal of the content from the UIArea as
implemented by the sap.ui.core.Popup no longer applies. The popup
removes the content only when it added it itself. As a result, a
destroy() of the dialog invalidated the static UIArea which in turn
destroyed other instances of the ViewSettingsDialog.

Change-Id: Iadfc2028bf5bb6f0b3ae2cc35767f43ca02a8543