[FIX] Configuration: fix test for language default in Chrome 40
Chrome version 40 seems to have changed the meaning of
'navigator.language'. It no longer reflects the users choice of
preferred languages (which since a few releases is available in the
'navigator.languages' (plural!) property), but instead returns the
installation or OS language.

Configuration.js had been adapted to the introduction of
'navigator.languages' some weeks ago, but the test case for the default
value of Configuration.js' language property had not. 

With Chrome 40 this mismatch now became obvious and made the test fail.

Change-Id: Ibd14d4ac9d744f3add7b78f59cef08c0b76fc254