[FEATURE] Templating: enabled aggregation helpers in controls helpers
The current implementation of UI5 Handlebars Templating did not allow to
use the aggregation helper to scope the aggregation to use for the
nested controls. Till now the aggregation helper could only be used for
the definition of new template controls. Now the following is possible:

{{#control id="event-table" sap-ui-type="sap.m.Table" inset="false"
items="{/items}"}}
  {{#aggregation name="columns"}}
    {{#control sap-ui-type="sap.m.Column"}}
      {{control sap-ui-type="sap.m.Text" text="Column"}}
    {{/control}}
  {{/aggregation}}
  {{#control sap-ui-type="sap.m.ColumnListItem"}}
    {{control sap-ui-type="sap.m.Text" text="{name}"}}
  {{/control}}
{{/control}}

The aggregation helper can define the aggregation which will be used for
the nested controls to add them to the parent control of the
aggregation. If not specifying the aggregation tag and the control has
no attribute "sap-ui-aggregation" then the child control will be
assigned to the default aggregation. 

Change-Id: Ie18e4acbddf063be678470586c9b406c7b8e65a0
Fixes  https://github.com/SAP/openui5/issues/270