[FIX] global padding mixin: remove !important with specific selectors
setting !important on the container padding was causing side effects
when using jQuery animations like "slideUp". This animates the paddings
of the div inline which was also overruled by the !important of the
padding concept.

We have removed the !important statement to not disturb animations and other padding effects. If another container uses this mixin it must make sure that the selector arguments are more specific than the other padding appliers.

Change-Id: Ia9e39a0cc4ebb796eb7f7fe0498fda9b13a9d5e3