[FIX] sap.m.Select: fix overlapping issue of the text (sap_bluecrystal theme)
In some scenarios when the width is set to less than 74px, the text
became overlapped with the arrow button.

The actual cause of the issue is a browser bug. See:

CSS text-overflow ellipsis is ignored (https://code.google.com/p/chromium/issues/detail?id=443159)

However changing the background color of the arrow on hover from rgba(222,222,222,0.5) to #f0f0f0 solve the issue.

Change-Id: I2a82add13a300591d6cb092c6664ce41cf0cb249
BCP: 1472016445