[FIX] sap.m.Select: fix overlapping issue of the text (sap_bluecrystal theme)
In some scenarios when the width is set to less than 74px, the text
became overlapped with the arrow button.

The actual cause of the issue is a browser bug. See:

CSS text-overflow ellipsis is ignored
(https://code.google.com/p/chromium/issues/detail?id=443159)

However changing the background color of the arrow solve the issue.

Change-Id: I9aa350d711c81293f0225bd2883165857ced8ba6
BCP: 1472016445
CR-Id: 012006153200001326462015