[INTERNAL] ComponentMetadata: finalized metadata/manifest merging
Fixed the declaration and accessor for the handleValidation property.

Fixed the component initialization to use the merged extension data in
order to make sure that the extensions configuration from the parent
component is also considered for the extensibility.

Support for mixing of component metadata and manifest. In case of
manifest entries exist the component metadata entries will be ignored. 

Added tests to validate the mixed declaration of metadata / manifest
incl. inheritance tests. 

Change-Id: Ic338bfec64da29978f31e583bde66335575f8945