[INTERNAL] includeStyleSheet: ignore search params of html page for href check
Make sure that includeStyleSheet is not including the URL parameters of the HTML 
page for the href check. This change makes sure that for creating the absolute URL
for the new link href the URL params of the HTML page are excluded. Otherwise it
is a strange behavior of URI.js that it considers the URL parameters of the base
URL when creating the absolute one except the relative URL itself also has URL 
parameters.

URI() => returns the URL of the HTML page incl. search string
URI().search("") => returns the URL of the HTML page excl. search string

URI(anyURL, baseUrl) => baseUrl must not incl. search string! 

Change-Id: Iefa1077af032e950040e489bf41681911dcecd36
CR-Id: 012006153200001936392015