[INTERNAL] Device API: fixed (Device|jQuery).support for PhantomJS
PhantomJS is no touch enabled browser even if the touch events are
exposed. So we need to make sure that the touch feature detection is
corrected properly.

Additionally make jquery.sap.global and jquery-mobile-custom reuse the
feature detection of the Device API to have a central place which can be
tweaked. 

Change-Id: If4b2f7829d2e6ab781855a1022f9e4a23931c7b9