[INTERNAL] QUnitTest: added dummy test for PhantomJS to avoid timeouts
In case of executing this QUnit Test with PhantomJS on windows the first
call of setTimeout will not be executed. This is somehow strange and
needs further investigations. The first assumption that the fakeTimers
of sinon are affecting this behavior was not correct. Even by extening
the ODataModelFakeService.js so that the setTimeout is followed by a
clock.tick call this will not lead to a proper execution of the test.
Only by adding a dummy function before it will now work properly.

Change-Id: I181fe329f371a9b39a28baddd21d14a8e4fc59df