[FIX] ManagedObject: proper cleanup of bindings
If applications use the eventing for bindings (dataRequested,
dataReceived), they can expect that after a control is destroyed (or
after a binding is removed), listeners attached to such bindings are no
longer active. This was not the case.

Object bindings (aka "Element" bindings) need to be cleaned up the same
way as other bindings when a ManagedObject is destroyed.

Change-Id: I198ee209070f78045889be661b2c2a402be610c1
CR-Id: 012006153200003631332015
BCP: 1570739341