[INTERNAL] PhantomJS: re-enabled focus tests due to available workaround
The :focus selector of jQuery is not working properly in combination
with PhantomJS. Therefore we needed to add a workaround into the
QUnitUtils to support the :focus selector again and to re-enable the
focus related tests. 

For the sap/m/Input.qunit.html the setTimeout and clearTimeout functions
needed to be overridden in order to make the test run reliable.
Otherwise strangewise the tests failed with the error "null is not a
constructor" when calling the destroy function of the input control.

Change-Id: I6c7408787b87828292c538b840ba16a13f2a31eb