[FIX] RenderManager: reset state before rendering (improves robustness)
When an exception occurred during rendering, the internal state of the 
RenderManager (esp. the list of rendered controls) was not cleaned up.
As a result, the next rendering called onAfterRendering also for 
the old content of the list (potentially calling controls that no longer
existed).

To avoid the penalty for an additinal "try catch", the state is just
cleaned up before rendering. 

The general contract of the Core that the behavior after exceptions 
is undefined and that exceptions must be avoided, is still valid!

Change-Id: I52479a933a6c23aa71348eae24eec3bde5522e1e