[INTERNAL][FIX] qunit-junit: add default module name
If no module was specified an empty string was used which resulted into
a dot at the end of the testsuite name in the JUnit report. This caused
issues when displaying the tests in Jenkins.
This change will make sure that the module name will not be empty.

Broken behavior was introduced with
76acf6580325ed4653a952ff7757d416652699d5

Change-Id: I3d240fa518d9ffd6ad2b297d81f2114e9559f659