[INTERNAL] md-template removed line item
This change removes the line item view, the line item not found view 
as well as everything related from the master detail template app.
It also touches the line items table in the detail view and the
corresponding functions in the controller in order to remove navigation
to the line item from there.

Change-Id: I313766aafd0374e9efe5431a46cecc008930e5e5