[FEATURE] sap.m.Popover: New placement types added
Four more placement types are added: VerticalPreferedTop,
VerticalPreferedBottom, HorizontalPreferedLeft, HorizontalPreferedRight.

For example VerticalPreferedTop tries to place the Popover at the Top side
of the opener if there is enough space. It consider enough space if the
Popover's height (for horizontal respectivly width) is less than the
available space on that side.

Change-Id: I592046d57a1c1e5f7617969118b944d96131a358