[FIX] Servlet: determine content type (c-t) for resources with c-t null
Firefox complains with an error message if a JSON or XML file is not
delivered with an appropriate content type. 

The servlet therefore should determine the content type by asking the
container. It already did so for resources where the internally used URL
connection returned a content type of 'content/unknown'. But it failed
to do so when 'null' was returned (e.g. for 'jndi:*' resources like
'jndi:/localhost/testsuite/resources/sap-ui-version.json')

Change-Id: I25a88815f898b962174714b3ae9716d5aab021ed