[FIX] core.Popup: De-Registration of Child Popups Fixed
- If a popup has child popups and the child popup was closed it wasn't
de-registered from the parent.
- Now if a popup is closed it checks if it was opened by another popup
(has got a parent) and de-registeres itself from its parent
- Struggled with the AutoComplete.qunit. During debugging I assumed that
a AutoComplete interferes another one. Therefore I restructured the
test. The reason why the tests failed was the new introduced check in
"Popup.checkDocking": Since the textbox of the AutoComplete was in the
top left corner the Popup closed automatically. By moving the textbox
the issue with the test was solved - but I kept the chances for the test
because I think it's way better than before and I didn't want to waste
my time... :-)

BCP: 1580000848
BCP: 1570149997
Change-Id: Ic4e8caad9181b6135a3681186ea8ef9fa624c85f