[FIX] ODataAnnotations: Multiple parameters in dynamic expressions
Dynamic expressions have not been correctly parsed before. Now they
are known to the parser and are added as arrays to the Annotation-
object.

The previously supported (but not "correct", according to the
now available Specification) behavior to allow multiple children for
Annotation elements has been kept but a warning is generated at 
runtime.

BCP: 1580027601
Change-Id: I384fa051930c452e874811b965f0adf0f21d8e52