[FEATURE] sap.m.OverflowToolbar - additional controls can move to the overflow
The action sheet is replaced by a popover, so that more controls have the possibility to move there.
Removed the offset previously taken by the arrow on Mobile.
Changed the overflowButton (from sap.m.Button to sap.m.ToggleButton)
Set the maxContentHeight of the popover to 60% (only on Mobile)
Restored the css width 100%
Removed Horizontal ScrollBar
Fixed ScrollBar bugs in mobile

Change-Id: Ibd69367209538ecb2d00a2b1cd19db87b42f65b6