[FIX] ux3.ToolPopup: ToolPopup Realizes When Content Rerenders
The ToolPopup (TP) has an own calculation of its size and position. If
content is added to the TP it will recalculate its size and position.
If the content of the TP rerenders due to a changed aggregation, the TP
didn't resize and re-position.

Now the TP registers a ResizeHandler to its DOM-reference and realizes
if the content width/height changed and a re-calculation of the width /
height or its position should be fixed.

BCP: 0020283765 0001000948 2015
CR-Id: 012003146900000986952015
Change-Id: I636170d94af208f90ee9c62cdbf9e888a8517997