[FIX] sinon.js: introduced workarounds for IE9 to fix QUnit issues
fakeXHR: readystatechanges errors need to be suppressed for IE9 and in
addition sinon need to respect the MSXML 3.0/6.0 status property for not
being completed. 

In case of <=IE9 UI5 enables the CORS support in jQuery to allow the
usage of jQuery.ajax function / sinon also needs to be synchronized with
this adoption by applying the CORS support flag from jQuery to sinon!

Change-Id: I94258d5552aa17501662f32f5bb23a31cdef01c4