[FIX] Make sap-ui-core-nojQuery.js work with JQueryFilter
When the JQueryFilter (a J2EE servlet filter that allows to override
UI5's default jQuery version) was used with the sap-ui-core.js version
from the OpenUI5 repository, it failed. The filter merges the desired
jQuery version with the sap-ui-core-nojQuery.js file on the fly. 
But the sap-ui-core-nojQuery.js from OpenUI5 fails when there is no
script tag with the same. 

With this fix, a script tag with 'sap-ui-core.js' in it is also
accepted.

The additinal change in testfwk.js helps to avoid future changes of that
file when UI5's default jQuery version changes.

Change-Id: I18bbb8ac3cda567ac0d9c009542e172ac7bc0174