[FIX] sap.m.Popover: Keep focus inside the Popover in Firefox
Keep the focus inside the Popover in firefox. The problem was that firefox
reports this firstfe and lastfe elements as visible and we set the focus
to them, instead to the real focusable element.

When we set the width and height to those two elements to 0 in the jQuery
check it will return that they are hidden (see jQuery.expr.filters.hidden
in jQuery).

Change-Id: I75232f5a26e33e3897f3933735ed212cead230fe
Fixes: https://github.com/SAP/openui5/issues/406
BCP: 1570151848