[INTERNAL] update eslint to 0.21.1
Fixed new eslint issues caused by bug fixes in rules.

Changed `consistent-this` rule from `error` to `warning` as there are a lot of new errors.
We need to check whether we want to fix the issues, stay with `warning` or contribute to eslint to allow the usage of `that` as function argument (optional rule parameter).

Change-Id: Ia4464acdf39a4f7e2ae9118e044d1e9c1030c4b4