[FIX] core.Control: Overwriting "_busyIndicatorDelayedCallId" in "onAfterRendering" Prevented
- When a control was initially set to busy the delayed call was
initialized in "onAfterRendering" as well. If the delay is still running
and a property was changed the control will be re-rendered as well and
the already running delayed call wasn't canceled.
"_busyIndicatorDelayedCallId" was simply overwritten. So both calls set
the control to busy and the control can't be set to busy=false again
- Now in "onAfterRendering" it's checked whether a delayed call is
running. If so the delay of this call will be used to show the busy
indicator animation.

BCP: 1570293283
BCP: 1570308931
BCP: 1570293457
BCP: 1570303316
CR-Id: 012003146900001819762015
Change-Id: Ia8462963cd7239e30e2cd4a07edf75b4c7823bbf