[FIX] v2.ODataModel: Instantiate annotations-object earlier
This change instantiates the annotations-object earlier in case it 
is known that it will be used for annotation parsing of the metadata
document later on. This way it is available to be used by the metamodel
independently of the time of its instantiation.

Also fixes the wrong visibility of the addAnnotation*-methods

Change-Id: I6447ed91e30b6b5577536ec861398c992aad1bbc