[FEATURE] JSONModel: Added observation mode
Usually when the application wants to change data in the JSONModel it
has to use setData() or setProperty() to make the changes known to the
model that the bindings can be updated. With ES5 we do now have a
possibility to enable observation of property changes on JSON data, so
when observation is activated, applications can just modify the JSON
data and binding will work automatically.

Change-Id: I9ba05bfab46d9547cef9d8e0b6777e15ec551da6