[FEATURE] sap.ui.core.Icon: add "useIconTooltip" property
This property allows to specify if the default Icon text should be
used as tooltip if there is no tooltip defined.

When the Icon control gets used standalone this is mostly the
expected behavior.
Controls which are using the Icon control internally might want
to decided wheater the tooltip should be rendered or not and should
set the flag accordingly.

TODO (see Icon.js#_getAccessibilityAttributes):
Decide whether the "tooltip" property should also override the
default text for "aria-label".

Change-Id: I424743d73fc1c32d690c3947ec4ba0fb1dedcad6