[FIX] sap.ui.core.Icon: remove default "title" based on icon-font
This is just a quick fix for the problem that controls like
sap.m.Button end up with 2 "title"s which are both read by
a screen-ready. This also prevents the Button's tooltip to show
up.

A better solution will be to let the control decide if the default
tooltip should be shown or that the control itself reads out the
icon's name and uses it as tooltip fallback.

Change-Id: If9b14eea17e118dab683f381e9ee13103e1fb907