[FIX] Popover: Passes the click event through the Popover arrow
In the past Popover arrow takes the click and if you click outside the
popover but near to the arrow the cklick event is taken from the arrow and
therefore the popover does not close itself.

Now we make the popover arrow to pass the click event and add a hidden
layer only below the arrow to collect the click events.

Change-Id: I4673ec08a9a52b24b86fd7d99d3446a79027b2e0
BCP: 1570317326