[FIX] BusyIndicator: BusyIndicator Attached to Popup's Open-Event
Since the BusyIndicator attaches itself as event delegate to the Popup
the BusyIndicator will be called with its event listener first. Because
the BusyIndicator's registration as delegate happens in "show". If the
BusyIndicator wouldn't register itself as event delegate, the Popup's
event will occur first - which could prevent the BusyIndicator from
display if a synchronous call is processed after calling the
BusyIndicator to show.

Change-Id: I0a471fb05e1e7fb6c3bd87859e321275916b2ecb