[INTERNAL] sap.m.semantic.SemanticPage: Fixed saving of current pageMode property
The current_pageMode variable used to be set as static valiable [to the
prototype of the SemanticPage], since is was considered as a global,
application-wide property. 

BCP incident 1570156830 however, shows a case where this property cannot
be savely shared and reused.

Fixed by redefining it as an instance property, belonging to the page
instance that has its pageMode.

Change-Id: I1f26683a8e7d219ac28c9b3a96db8858a5738e45