[FIX] sap.m.Popover: Correctly show ResponsiveGridLayout when visibility is binded
When use ResponsiveGridLayout in Popover and the Grid is initially set
with visibility false, and after that the visibility is changed to true,
the Grid is not shown. This is due to the fact, that it trys to get %
width of the parent, but parent width initially is set to 0.

Change-Id: I22e89737d1da802fd4e69e850983a70efe6c1ccb
CR-Id: 012006153200002063752015
BCP: 1570454730