[FIX] sap.m.Popover: beforeClose event was fired but Popover was not closed
When you have sap.m.Popover opened inside another sap.m.Popover and move
the focus inside the parent sap.m.Popover but outside of the child
sap.m.Popover, the autoClose mechanism of the second is triggered.

Inside the onFocusEvent of the child core.Popup, autoClose mechanism send the
blur event throught the EventBus to the parent core.Popup to check if the
Focus is still inside the Parent core.Popup.

Due to the fact that the sap.m.Popover overrides the core.Popup close
method, the beforeClose event is fired, after that core.Popup close method
is called and it checks if the Popup has still has the focus. And if this
is the case the Popup is not closed, but the beforeClose event is fired.

Change-Id: I63d35d946c4710f57beb40fa954729cf813959eb
BCP: 1570460654
CR-Id: 012006153200002169062015