[FIX] sap.ui.model.Filter: String normalization clientside
For the assurance of correct filtering, all data gets normalized when
comparing and normalization is available. This means canonical
equivalent strings compare as equal per default for our Filter in all
desktop browsers and mobile browsers capable of
String.prototype.normalize. The applied normlization for is "NFC" as
recommended by W3C.

http://www.w3.org/TR/2012/WD-charmod-norm-20120501/#sec-Normalization

Change-Id: I6a54a77f3aace3bfe9ab4cda0aeabbaf2a813de7