[INTERNAL] Support delta compression in CLDR files.
Region specific CLDR files only need to contain the delta to their
corresponding region-independent CLDR file. This compression so far 
will only be applied in the sap.fiori bundle where all CLDR files 
for a language are contained in the same bundle. The CLDR files in the 
sap.ui.core are not touched to avoid an additional roundtrip.

Change-Id: Ia6e9a829a05bc28509fb1b31f2f4b6a932d2a11c