[INTERNAL] Allow to preload a component without 'requiring' it.
This allows to trigger the preload of a component in parallel to 
the loading of other preload bundles and to require the Component
controller module only after both tasks have finished.

Additionally, the registration of URL paths is now executed even when
the component preload is disabled (helpful in debug mode). 

Last but not least, asyncHints now can list additional preload bundles
that should be loaded in parallel.

Change-Id: Ie8c9b850bdfbb628a8e1e4b097646e17bb287539