[FIX] Anonymous define in URI.js breaks CVOM charts
Inside WebDynpro ABAP, the VIZ charting library is integrated by loading
some of the sap.viz native scripts. One of these scripts contains a copy
of requireJS, so requireJS and its define() API are present in the page.

When an WD ABAP HTML island then loads UI5 after the VIZ library, 
the URI.js embedded in sap-ui-core.js recognizes the existence of 
define() and tries to register with it as an anonymous module.

But the current version of requireJS rejects this if it didn't load the
module itself (which is not the case).

As a workaround, URI.js has been patched to call define with the module
name that matches the UI5 default location of URI.js
(sap/ui/thirdparty/URI). This is an inexpensive but not 100% perfect
solution as it not only affects the URI.js embedded in sap-ui-core.js
but any usage of URI.js from the UI5 resource pool.

However, the risk that URI.j is used from the UI5 resource pool but 
with a different name was regarded quite low. Should an app really need
a different name, it might embed the original URI.js code, use requireJS
configuration or create a wrapper module that requires URI.js with the
UI5 name and exposes it wit hthe desired name.

(cherry picked from commit 3a9e6ea70aaa9831a28da92f833d28bed6fafeed)

Change-Id: Ib13e928ef4e2aa3a345dbfe3524d4a4fa4230d26
BCP: 1570526495
BCP: 1570447997
CR-Id: 012006153200002834302015