[FEATURE] Component: enabled sap.ui5/resourceRoots / minUI5Version warn
The resource roots which are maintained in the manifest.json underneath
the section sap.ui5/resourceRoots will be registered automatically by
the ComponentMetadata once it is being created. Right now the resource
roots will not be considered for the sap.ui.define dependencies. This
will be implemented as part of the manifest first approach.

In addition the sap.ui5/dependencies/minUI5Version will report a warning
in case of the log level warning is enabled. This will prevent the load
of the sap-ui-version.json in case of normal UI5 app execution where
only the log level error is enabled. This is important because the
sap-ui-version.json is a sync request.

Change-Id: If05f8a78ffdf3b61389aed5f361c963a841e540e