[FIX] ux3.ToolPopup: Remove DOM when hiding ToolPopup
Before the DOM was only hidden via CSS, which could lead to controls
being rendered twice (and thus duplicate IDs in the dom). As the 
ToolPopup can only contain controls and not arbitrary HTML like the 
Popup itself, it is safe to remove the DOM entirely.

See also change with IDs
 I92b1f2aafa5988c1e46e37c7b4ece87d90cb4c62 and
 I612e1c76409ca91c4a0ac35c6e1d6eb1f16bcb3a

CR-Id: 012006153200002805822015
Change-Id: I47fa1efbcfde1d22a5d345a7084cd49e93071ec9